requirements.txt
Django==2.2.10
gunicorn==19.7.1
whitenoise==3.3.1
django-allauth==0.40.0
django-extensions==2.2.9


astroid==2.3.3
defusedxml==0.5.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
oauthlib==3.1.0
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
typed-ast==1.4.1
coverage==5.1
django-nested-admin==3.3.0
nose==1.3.7
six==1.14.0
nltk==3.5
pandas==1.0.3
numpy==1.18.3
vaderSentiment==3.3.1
matplotlib==3.2.1
textblob==0.15.3
